fix: fixing reset password api
