+++ src/routes/auth.js
@@ -2,8 +2,10 @@ import { Router } from 'express';
 import * as userController from '../controllers/user.js';
 import * as authController from '../controllers/auth.js';
 import * as otpController from '../controllers/otp.js';
+import * as passwordResetController from '../controllers/passwordReset.js';
 import * as userValidationMiddleware from '../middlewares/validation/user.js';
 import * as authValidationMiddleware from '../middlewares/validation/auth.js';
+import * as passwordResetValidationMiddleware from '../middlewares/validation/passwordReset.js';
 
 export default (app) => {
   const router = Router();
@@ -28,18 +30,19 @@ export default (app) => {
 
   router.post(
     '/reset-password/request',
-    authValidationMiddleware.resetPasswordRequestValidation,
-    authController.sendResetPasswordEmail
+    passwordResetValidationMiddleware.resetPasswordRequestValidation,
+    passwordResetController.sendResetPasswordEmail
   );
 
   router.get(
     '/reset-password/validate/:token',
-    authController.validateResetPasswordToken
+    passwordResetValidationMiddleware.validateTokenParams,
+    passwordResetController.validateResetPasswordToken
   );
 
   router.post(
     '/reset-password',
-    authValidationMiddleware.resetPasswordValidation,
-    authController.resetPassword
+    passwordResetValidationMiddleware.resetPasswordValidation,
+    passwordResetController.resetPassword
   );
 };
